Add direction flag for UFFDIO_REMAP
The direction flag in the remap struct indicates which vma list to
search. This is needed for the use case of evicting a page from
outside the context of the userfaultfd. For example: eviction from
a userfault vma in userfault_ctx->mm to a vma in current->mm,
requires that we validated the src page exists in the vma list of
userfault->mm, where move a page into the userfaultfd context
requires that the src page must exist in the vma list of current->mm
and that the destination page exist in the vma list of
userfault_ctx->mm.