Fix SIGBUS when a page fault is retried
Note: this only was quick fix so that code would run. Did not look for
the complete solution.

Sometimes handle_userfault gets a fault that is on its second try. It
is reasonable that handle_userfault will allow another retry because
this can be the case when page was swapped out. This patch prevents
handle_userfault from returing SIGBUS when such as case occurs.