userfaultfd: non-cooperative: Add fork() event
When the mm with uffd-ed vmas fork()-s the respective vmas notify their
uffds with the event which contains a descriptor with new uffd.  This new
descriptor can then be used to get events from the child and populate its
mm with data.  Note, that there can be different uffd-s controlling
different vmas within one mm, so first we should collect all those uffds
(and ctx-s) in a list and then notify them all one by one but only once
per fork().

The context is created at fork() time but the descriptor, file struct and
anon inode object is created at event read time.  So some trickery is
added to the userfaultfd_ctx_read() to handle the ctx queues' locking vs
file creation.

Another thing worth noticing is that the task that fork()-s waits for the
uffd event to get processed WITHOUT the mmap sem.

Link: http://lkml.kernel.org/r/20161216144821.5183-9-aarcange@redhat.com
Signed-off-by: Pavel Emelyanov <xemul@parallels.com>
Signed-off-by: Mike Rapoport <rppt@linux.vnet.ibm.com>
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Cc: "Dr. David Alan Gilbert" <dgilbert@redhat.com>
Cc: Hillf Danton <hillf.zj@alibaba-inc.com>
Cc: Michael Rapoport <RAPOPORT@il.ibm.com>
Cc: Mike Kravetz <mike.kravetz@oracle.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>