mm/zsmalloc: remove redundant SetPagePrivate2 in create_page_chain
We had used page->lru to link the component pages (except the first page)
of a zspage, and used INIT_LIST_HEAD(&page->lru) to init it.  Therefore,
to get the last page's next page, which is NULL, we had to use page flag
PG_Private_2 to identify it.

But now, we use page->freelist to link all of the pages in zspage and init
the page->freelist as NULL for last page, so no need to use PG_Private_2
anymore.

This remove redundant SetPagePrivate2 in create_page_chain and
ClearPagePrivate2 in reset_page().  Save a few cycles for migration of
zsmalloc page :)

Link: http://lkml.kernel.org/r/1487076509-49270-1-git-send-email-xieyisheng1@huawei.com
Signed-off-by: Yisheng Xie <xieyisheng1@huawei.com>
Reviewed-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>
Acked-by: Minchan Kim <minchan@kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>