userfaultfd: UFFDIO_REMAP
This remap ioctl allows to atomically move a page out of a userfaultfd
address space. It's more expensive than "copy" as it requires a TLB
flush on the source range for each ioctl, which is an expensive
operation on SMP. For this reason, "remap" "out" is the only direction
supported, where movement into a userfaultfd address space can be
achieved using "copy". This ioctl can be used non-cooperatively by a
user space process to limit the number of pages mapped in a
userfaultfd address space.

Co-authored-by: Blake Caldwell <blake.caldwell@colorado.edu>
Signed-off-by: Blake Caldwell <blake.caldwell@colorado.edu>