userfaultfd: add direction flag for UFFDIO_REMAP
This is needed to support the two use cases, of moving a page from
a userland anonymous vma into the userfaultfd registered region, or
vice-versa. Moving a page from the userfaultfd region is necessary
to remove pages, and generate page fault events if the application
tries accessing the page again.